requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
ortools==9.8.3296
firebase-admin==6.4.0
pydantic==2.6.0
python-dotenv==1.0.1
google-genai==0.3.0